Fixed all warnings that appeared with DEBUG=1 set. Also fixed missing return codes in BufferPool::ReleaseBuffer() and Journal::WriteLogEntry().
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@2845 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/add-ons/kernel/file_systems/bfs/BPlusTree.cpp
@@ -1613,7 +1613,7 @@ BPlusTree::Find(const uint8 *key, uint16 keyLength, off_t *_value)
 				*_value = node->Values()[keyIndex];
 
 #ifdef DEBUG
-			if (levels != fHeader->max_number_of_levels)
+			if (levels != (int32)fHeader->max_number_of_levels)
 				DEBUGGER(("levels don't match"));
 #endif
 			return status;
